Fix changing of cpp build configuration
Previously, to change build configuration, we were relying on the fact
that calling "this.restart()" in onActiveBuildConfigChanged would cause
a new LanguageClient to be instantiated.  This would call
createOptions(), where we would return new value of
compilationDatabasePath in the initializationOptions that are sent to
clangd.

However, since commit bfb235a0d7b ("[languages] fix leaking language
clients"), this.restart() does not create a new client, it re-uses the
existing one.  createOptions is not called and the old setting of
compilationDatabasePath is sent to clangd.  Therefore, the new build
configuration does not take effect.

This patch fixes it locally in the cpp package: we now go overwrite the
initializationOptions in the existing LanguageClient object, so that the
new value will be picked up when call restart.

Refs #4282

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>